requirements.txt
# requirements.txt
pandas
numpy  
scikit-learn
matplotlib
seaborn
xgboost
fastapi
uvicorn
jupyter
python-dotenv
requests
plotly
joblib